fix: wasm build
